Fix message disposing when session is closed or not initialized
