+++ src/Nethermind/Nethermind.Network.Test/P2P/PacketSenderTests.cs
@@ -38,7 +38,6 @@ namespace Nethermind.Network.Test.P2P
             PacketSender packetSender = new(serializer, LimboLogs.Instance, TimeSpan.Zero);
             packetSender.HandlerAdded(context);
             packetSender.Enqueue(testMessage);
-            testMessage.WasDisposed.Should().BeTrue();
 
             context.Received(1).WriteAndFlushAsync(Arg.Any<IByteBuffer>());
         }
@@ -62,7 +61,6 @@ namespace Nethermind.Network.Test.P2P
             PacketSender packetSender = new(serializer, LimboLogs.Instance, TimeSpan.Zero);
             packetSender.HandlerAdded(context);
             packetSender.Enqueue(testMessage);
-            testMessage.WasDisposed.Should().BeTrue();
 
             context.Received(0).WriteAndFlushAsync(Arg.Any<IByteBuffer>());
         }
@@ -88,7 +86,6 @@ namespace Nethermind.Network.Test.P2P
             PacketSender packetSender = new(serializer, LimboLogs.Instance, delay);
             packetSender.HandlerAdded(context);
             packetSender.Enqueue(testMessage);
-            testMessage.WasDisposed.Should().BeTrue();
 
             await context.Received(0).WriteAndFlushAsync(Arg.Any<IByteBuffer>());
 
@@ -101,13 +98,6 @@ namespace Nethermind.Network.Test.P2P
         {
             public override int PacketType { get; } = 0;
             public override string Protocol { get; } = "";
-
-            public bool WasDisposed { get; set; }
-            public override void Dispose()
-            {
-                base.Dispose();
-                WasDisposed = true;
-            }
         }
     }
 }
